// src/services/api/project/project.ts
export interface Project {
  title: string;
  shortTitle: string;
  description: string;
  startDate: string;
  publishDate: string;
  technoList: string[];
  technoElements: string[];
  logoUri: string;
  clientInformation: string;
  link: string;
  version: string;
  status: string;
  type: string;
  priority: string;
}